Feature/top up business logic (#7)
* create top up bloc with add beneficiary logic

* fix user repository

* refactor transactions and beneficiaries repositories and models

* create top up business logic

+++ packages/user_repository/lib/src/user_repository.dart
@@ -1,5 +1,5 @@
 abstract class UserRepository {
-  Future<void> addCreditToBalance(String id, int amount);
+  Future<void> addCreditToBalance(String id, double amount);
 
-  Future<void> removeCreditFromBalance(String id, int amount);
+  Future<void> removeCreditFromBalance(String id, double amount);
 }
